Implemented No Share, Cheapest and Cheapest Users energy consumption policies
Refactor main into two parts for easier navigation.
Fixed constant AP and User lists.

The same AP list is passed into simulations. An improvement would be to
make a copy of the list for each simulation sequence so that everything
else is made constant.

TODO: Fix plots for relationship graphs