Modified Energy Consumption Rate
Refactored the main function which calls testPolicy() to run simulations
related to energy policy testing.

A deepcopy of the Access Points and Users are now created in every
simulation for consistency.

Init now creates one usrlist with randint and another with ppp. Both are
returned on the function call for simplicity.